Login/Registration Logic update
Login/Registration Logic update
